Introduce `Command` Domain Object
This PR also introduces a `Container` class that defines most of the functionality
required by Protean data containers.
