+++ src/protean/domain.py
@@ -32,6 +32,7 @@ _sentinel = object()
 class DomainObjects(Enum):
     AGGREGATE = 'AGGREGATE'
     APPLICATION_SERVICE = 'APPLICATION_SERVICE'
+    COMMAND = 'COMMAND'
     DATA_TRANSFER_OBJECT = 'DATA_TRANSFER_OBJECT'
     DOMAIN_EVENT = 'DOMAIN_EVENT'
     DOMAIN_SERVICE = 'DOMAIN_SERVICE'
@@ -111,6 +112,7 @@ class Domain(_PackageBoundObject):
     from protean.core.aggregate import BaseAggregate
     from protean.core.application_service import BaseApplicationService
     from protean.core.broker.subscriber import BaseSubscriber
+    from protean.core.command import BaseCommand
     from protean.core.data_transfer_object import BaseDataTransferObject
     from protean.core.domain_event import BaseDomainEvent
     from protean.core.domain_service import BaseDomainService
@@ -143,6 +145,7 @@ class Domain(_PackageBoundObject):
     base_class_mapping = {
             DomainObjects.AGGREGATE.value: BaseAggregate,
             DomainObjects.APPLICATION_SERVICE.value: BaseApplicationService,
+            DomainObjects.COMMAND.value: BaseCommand,
             DomainObjects.DATA_TRANSFER_OBJECT.value: BaseDataTransferObject,
             DomainObjects.DOMAIN_EVENT.value: BaseDomainEvent,
             DomainObjects.DOMAIN_SERVICE.value: BaseDomainService,
@@ -257,6 +260,10 @@ class Domain(_PackageBoundObject):
     @property
     def application_services(self):
         return self._domain_registry._elements[DomainObjects.APPLICATION_SERVICE.value]
+
+    @property
+    def commands(self):
+        return self._domain_registry._elements[DomainObjects.COMMAND.value]
 
     @property
     def data_transfer_objects(self):
@@ -447,6 +454,11 @@ class Domain(_PackageBoundObject):
     def application_service(self, _cls=None, aggregate_cls=None, bounded_context=None, **kwargs):
         return self._domain_element(
             DomainObjects.APPLICATION_SERVICE, _cls=_cls, **kwargs,
+            aggregate_cls=aggregate_cls, bounded_context=bounded_context)
+
+    def command(self, _cls=None, aggregate_cls=None, bounded_context=None, **kwargs):
+        return self._domain_element(
+            DomainObjects.COMMAND, _cls=_cls, **kwargs,
             aggregate_cls=aggregate_cls, bounded_context=bounded_context)
 
     def data_transfer_object(self, _cls=None, aggregate_cls=None, bounded_context=None, **kwargs):
